Create AmicableNumbers.java
Developed the following methods: method to compute and return the sum of the proper divisors of an integer num parameter, method to return the boolean state of two integer parameters being an amicable number pair, and method to compute and return the sum of amicable numbers from 1 to an integer limit parameter, taking advantage of a HashSet. All implemented in main method.